Move tabs into separate components.
