+++ src/components/App.tsx
@@ -1,30 +1,19 @@
 import { ChakraProvider, Tab, TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react";
-import React, { useEffect, useState } from "react";
 import {
-  createDataContext,
-  createItems,
-  createNewCollection,
-  createTable,
-  getAllItems,
-  getDataContext,
   initializePlugin,
   addDataContextChangeListener,
   ClientNotification,
 } from "@concord-consortium/codap-plugin-api";
+import React, { useEffect, useState } from "react";
+import {
+  kAboutTabLabel, kDataContextName, kGraphTabLabel, kInitialDimensions, kPluginName, kVersion
+} from "../utilities/constants";
+import { AboutTab } from "./about-tab";
 import "./App.css";
-
-const kPluginName = "Space-Time Cube";
-const kVersion = "0.0.1";
-const kInitialDimensions = {
-  width: 380,
-  height: 680
-};
-const kDataContextName = "SamplePluginData";
+import { GraphTab } from "./graph-tab";
 
 export const App = () => {
-  const [codapResponse, setCodapResponse] = useState<any>(undefined);
   const [listenerNotification, setListenerNotification] = useState<string>();
-  const [dataContext, setDataContext] = useState<any>(null);
 
   useEffect(() => {
     initializePlugin({pluginName: kPluginName, version: kVersion, dimensions: kInitialDimensions});
@@ -40,85 +29,24 @@ export const App = () => {
     addDataContextChangeListener(kDataContextName, casesUpdatedListener);
   }, []);
 
-  const handleOpenTable = async () => {
-    const res = await createTable(kDataContextName);
-    setCodapResponse(res);
-  };
-
-  const handleCreateData = async() => {
-    const existingDataContext = await getDataContext(kDataContextName);
-    let createDC, createNC, createI;
-    if (!existingDataContext.success) {
-      createDC = await createDataContext(kDataContextName);
-      setDataContext(createDC.values);
-    }
-    if (existingDataContext?.success || createDC?.success) {
-      createNC = await createNewCollection(kDataContextName, "Pets", [
-        { name: "animal", type: "categorical" },
-        { name: "count", type: "numeric" }
-      ]);
-      createI = await createItems(kDataContextName, [
-        { animal: "dog", count: 5 },
-        { animal: "cat", count: 4 },
-        { animal: "fish", count: 20 },
-        { animal: "horse", count: 1 },
-        { animal: "bird", count: 2 },
-        { animal: "snake", count: 1 }
-      ]);
-    }
-
-    setCodapResponse(`
-      Data context created: ${JSON.stringify(createDC)}
-      New collection created: ${JSON.stringify(createNC)}
-      New items created: ${JSON.stringify(createI)}
-    `);
-  };
-
-  const handleGetResponse = async () => {
-    const result = await getAllItems(kDataContextName);
-    setCodapResponse(result);
-  };
-
   return (
     <ChakraProvider>
       <div className="App">
         <Tabs variant="enclosed">
           <TabList>
             <Tab>
-              3D Graph
+              {kGraphTabLabel}
             </Tab>
             <Tab>
-              About
+              {kAboutTabLabel}
             </Tab>
           </TabList>
           <TabPanels>
             <TabPanel>
-              <div className="buttons">
-                <button onClick={handleCreateData}>
-                  Create some data
-                </button>
-                <button onClick={handleOpenTable} disabled={!dataContext}>
-                  Open Table
-                </button>
-                <button onClick={handleGetResponse}>
-                  See getAllItems response
-                </button>
-                <div className="response-area">
-                  <span>Response:</span>
-                  <div className="response">
-                    {codapResponse && `${JSON.stringify(codapResponse, null, "  ")}`}
-                  </div>
-                </div>
-              </div>
-              <div className="response-area">
-                  <span>Listener Notification:</span>
-                  <div className="response">
-                    {listenerNotification && listenerNotification}
-                  </div>
-              </div>
+              <GraphTab listenerNotification={listenerNotification} />
             </TabPanel>
             <TabPanel>
-              About the Space-Time Cube...
+              <AboutTab />
             </TabPanel>
           </TabPanels>
         </Tabs>
