+++ src/components/about-tab.tsx
@@ -1,0 +1,3 @@
+export function AboutTab() {
+  return "About the Space-Time Cube...";
+}
