+++ src/utilities/constants.ts
@@ -1,0 +1,10 @@
+export const kPluginName = "Space-Time Cube";
+export const kVersion = "0.0.1";
+export const kInitialDimensions = {
+  width: 380,
+  height: 680
+};
+export const kDataContextName = "SamplePluginData";
+
+export const kGraphTabLabel = "3D Graph";
+export const kAboutTabLabel = "About";
